requirements.txt
streamlit
google-generativeai
transformers
torch
Pillow
moviepy
colorthief
requests
